Breaking game logic onto useTcMatchRunner and updating TwoColumnMatch component
